+++ CHANGES.md
@@ -4,6 +4,7 @@
     * `to_float` template function
     * `PathBuf`s can now be used as templates
     * BUGFIX: tick events no longer tries to use removed components
+    * BUGFIX: erasing characters correctly between frames
 * 0.2.5
     * BUGFIX: component reuse in if / else
     * `with` statement