+++ CHANGES.md
@@ -1,3 +1,10 @@
+* 0.2.6
+    * `Either` now works by doing truthiness checks on state
+    * `Backend::full_screen()` convenience function
+    * `to_float` template function
+    * `PathBuf`s can now be used as templates
+    * BUGFIX: tick events no longer tries to use removed components
+    * BUGFIX: erasing characters correctly between frames
 * 0.2.5
     * BUGFIX: component reuse in if / else
     * `with` statement